mod graphic and math

+++ project/rcBase/src/rc_types.h
@@ -11,10 +11,10 @@
 
 typedef char           s8;
 typedef short          s16;
-typedef long           s32;
+typedef int            s32;
 typedef unsigned char  u8;
 typedef unsigned short u16;
-typedef unsigned long  u32;
+typedef unsigned int   u32;
 typedef float          f32;
 typedef double         f64;
